ER-701 retrieve lists only once per execution of rules
