+++ src/Shared/Recruit.Vacancies.Client/Application/Rules/BaseRules/ConsolidationOption.cs
@@ -1,0 +1,8 @@
+﻿namespace Esfa.Recruit.Vacancies.Client.Application.Rules.BaseRules
+{
+    public enum ConsolidationOption
+    {
+        NoConsolidation,
+        ConsolidateByField
+    }   
+}
