+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/Mongo/MongoDbConventions.cs
@@ -18,6 +18,7 @@ namespace Esfa.Recruit.Vacancies.Client.Infrastructure.Mongo
             ConventionRegistry.Register("recruit conventions", pack, t => true);
 
             BsonClassMap.RegisterClassMap<ProfanityData>();
+            BsonClassMap.RegisterClassMap<BannedPhrasesData>();
         }
     }
 }
